:lipstick: fix: change tabContent #22
