+++ frontend/src/component/Carousel.tsx
@@ -1,36 +1,53 @@
-import LentModal from '../modal/LentModal'
+import CabinetBox from './CabinetBox'
+import './carousel.css'
+import { cabinetInfo } from '../pages/Lent'
 
 export default function Carousel(props:any){
-    return (
-        <div id="carouselExampleIndicators" class="carousel slide" data-bs-ride="carousel">
-            
-            <div class="carousel-indicators">
-              <button type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide-to="0" class="active" aria-current="true" aria-label="Slide 1"></button>
-              <button type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide-to="1" aria-label="Slide 2"></button>
-              <button type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide-to="2" aria-label="Slide 3"></button>
-            </div>
+  const cabinetBlock = (f_idx:number, s_idx:number) => {
+    if (!props.info || !props.info.cabinet)
+      return [];
+    let list = [];
+    const cab:Array<cabinetInfo> = props.info.cabinet[props.l_idx][f_idx][s_idx];
+    for (let i = 0; i < cab.length; i++){
+      const id= props.outer_lent.find((l:any)=>l.lent_cabinet_id === cab[i].cabinet_id);
+      list.push(
+        <CabinetBox className="d-block w-100" key={`cab_box_${cab[i].cabinet_id}`} cabinet_id={cab[i].cabinet_num} intra_id={id ? id.intra_id : ""}></CabinetBox>
+      );
+    }
+    return list;
+  }
 
-            <div class="carousel-inner">
-              <div class="carousel-item active">
-                <img src="..." class="d-block w-100" alt="...">
-              </div>
-              <div class="carousel-item">
-                <img src="..." class="d-block w-100" alt="...">
-              </div>
-              <div class="carousel-item">
-                <img src="..." class="d-block w-100" alt="...">
-              </div>
-            </div>
-
-            
-            <button class="carousel-control-prev" type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide="prev">
-              <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-              <span class="visually-hidden">Previous</span>
-            </button>
-            <button class="carousel-control-next" type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide="next">
-              <span class="carousel-control-next-icon" aria-hidden="true"></span>
-              <span class="visually-hidden">Next</span>
-            </button>
+  const navSection = (idx:number) => {
+    let list = [];
+    if (!props.info || !props.info.section || props.info.section[props.l_idx].length <= idx)
+      return [];
+    console.log(props.info.section[props.l_idx][idx].length);
+    for (let i = 0; i < props.info.section[props.l_idx][idx].length; i++){
+      list.push(
+        <div className={`carousel-item${i ? '' : ' active'}`} key={`carousel-item_${props.info.section[props.l_idx][idx][i]}`} >
+          <label className="m-3" key={`label_${props.info.section[props.l_idx][idx][i]}`}>{props.info.section[props.l_idx][idx][i]}</label>
+          {cabinetBlock(idx, i)}
         </div>
-    );
+      );
+    }
+    return list;
+  }
+
+  return (
+    <div className={`tab-pane${props.outer_i ? '' : ' active'}`} id={`nav-${props.floor_name}`} key={`nav-${props.floor_name}`} role="tabpanel" aria-labelledby={`nav-${props.floor_name}-tab`}>
+      <div id={`carousel_${props.l_idx}_${props.floor_name}`} className="carousel slide" data-bs-ride="carousel">
+        <div className="carousel-inner" key={`nav-inner${props.floor_name}`}>
+          {navSection(props.outer_i)}
+        </div>
+        <button className="carousel-control-prev" type="button" data-bs-target={`#carousel_${props.l_idx}_${props.floor_name}`} data-bs-slide="prev">
+          <span className="carousel-control-prev-icon" aria-hidden="true"></span>
+          <span className="visually-hidden">Previous</span>
+        </button>
+        <button className="carousel-control-next" type="button" data-bs-target={`#carousel_${props.l_idx}_${props.floor_name}`} data-bs-slide="next">
+          <span className="carousel-control-next-icon" aria-hidden="true"></span>
+          <span className="visually-hidden">Next</span>
+        </button>
+      </div>
+    </div>
+  );
 }
