+++ frontend/src/pages/Lent.tsx
@@ -2,10 +2,11 @@ import {useEffect, useState} from 'react'
 import axios from 'axios'
 import Location from '../component/Location'
 import Menu from '../component/Menu'
-// import LentModal from '../modal/LentModal'
+import LentModal from '../modal/LentModal'
 import './lent.css'
 import './main.css'
 import CabinetBox from '../component/CabinetBox'
+import Carousel from '../component/Carousel'
 
 export type cabinetInfo = {
   cabinet_id: number,
@@ -31,7 +32,6 @@ type lentInfo = {
   extension: boolean,
   intra_id?: string
 }
-
 
 export default function Lent(){
     const [l_idx, setLidx] = useState<number>(0);
@@ -78,41 +78,6 @@ export default function Lent(){
       }
       return list;
     }
-    // const navContent = () => {
-    //   let list = [];
-    //   if (!info || !info.floor || info.floor.length <= l_idx)
-    //     return [];
-    //   for (let i = 0; i < info.floor[l_idx].length; i++){
-    //     let floor_name = info.floor[l_idx][i];
-    //     list.push(
-    //       <div className={`tab-pane${i ? '' : ' active'}`} id={`nav-${floor_name}`} key={`nav-${floor_name}`} role="tabpanel" aria-labelledby={`nav-${floor_name}-tab`}>
-    //         {navSection(i)}
-    //       </div>
-    //     );
-    //   }
-    //   return list;
-    // }
-
-    // const navSection = (idx:number) => {
-    //   let list = [];
-    //   if (!info || !info.section || info.section[l_idx].length <= idx)
-    //     return [];
-    //   // console.log(info.section[l_idx][idx].length);
-    //   for (let i = 0; i < info.section[l_idx][idx].length; i++){
-    //     list.push(
-    //       <div key={`nav_section_${info.section[l_idx][idx][i]}`}>
-    //         <label className="m-3" key={`label_${info.section[l_idx][idx][i]}`}>{info.section[l_idx][idx][i]}</label>
-    //         <div className="row mx-1" key={`block_${info.section[l_idx][idx][i]}`}>
-    //           {cabinetBlock(idx, i)}
-    //         </div>
-    //       </div>
-    //     );
-    //   }
-    //   return list;
-    // }
-
-
-    // 상단의 두 컴포넌트 코드가 carousel 적용시도 전 코드입니다! 이하 코드는 적용 중 코드!
     const navContent = () => {
       let list = [];
       if (!info || !info.floor || info.floor.length <= l_idx)
@@ -120,63 +85,7 @@ export default function Lent(){
       for (let i = 0; i < info.floor[l_idx].length; i++){
         let floor_name = info.floor[l_idx][i];
         list.push(
-          // <div className={`tab-pane${i ? '' : ' active'}`} id={`nav-${floor_name}`} key={`nav-${floor_name}`} role="tabpanel" aria-labelledby={`nav-${floor_name}-tab`}>
-          //   {navSection(i)}
-          // </div>
-          <div className={`tab-pane${i ? '' : ' active'}`} 
-          id={`nav-${floor_name}`} key={`nav-${floor_name}`} 
-          role="tabpanel" aria-labelledby={`nav-${floor_name}-tab`}>
-            <div id={`carousel${l_idx}_${floor_name}`} className="carousel slide" data-bs-ride="carousel">
-              <div className="carousel-inner" key={`nav-inner${floor_name}`}>
-                {navSection(i)}
-              </div>
-              <button className="carousel-control-prev" type="button" data-bs-target={`#carousel${l_idx}_${floor_name}`} data-bs-slide="prev">
-                <span className="carousel-control-prev-icon" aria-hidden="true"></span>
-                <span className="visually-hidden">Previous</span>
-              </button>
-              <button className="carousel-control-next" type="button" data-bs-target={`#carousel${l_idx}_${floor_name}`} data-bs-slide="next">
-                <span className="carousel-control-next-icon" aria-hidden="true"></span>
-                <span className="visually-hidden">Next</span>
-              </button>
-            </div>
-          </div>
-        );
-      }
-      return list;
-    }
-
-    const navSection = (idx:number) => {
-      let list = [];
-      if (!info || !info.section || info.section[l_idx].length <= idx)
-        return [];
-      // console.log(info.section[l_idx][idx].length);
-      for (let i = 0; i < info.section[l_idx][idx].length; i++){
-        list.push(
-          // <div key={`nav_section_${info.section[l_idx][idx][i]}`}>
-          //   <label className="m-3" key={`label_${info.section[l_idx][idx][i]}`}>{info.section[l_idx][idx][i]}</label>
-          //   <div className="row mx-1" key={`block_${info.section[l_idx][idx][i]}`}>
-          //     {cabinetBlock(idx, i)}
-          //   </div>
-          // </div>
-          <div className={`carousel-item${i ? '' : ' active'}`}>
-            {cabinetBlock(idx, i)}
-          </div>
-
-        );
-      }
-      return list;
-    }
-
-    // 여기까지가 carousel 적용시도 중 코드입니다!
-    const cabinetBlock = (f_idx:number, s_idx:number) => {
-      if (!info || !info.cabinet)
-        return [];
-      let list = [];
-      const cab:Array<cabinetInfo> = info.cabinet[l_idx][f_idx][s_idx];
-      for (let i = 0; i < cab.length; i++){
-        const id= lent.find((l)=>l.lent_cabinet_id === cab[i].cabinet_id);
-        list.push(
-          <CabinetBox className="d-block w-100" key={`cab_box_${cab[i].cabinet_id}`} cabinet_id={cab[i].cabinet_num} intra_id={id ? id.intra_id : ""}></CabinetBox>
+          <Carousel info={info} l_idx={l_idx} outer_i={i} outer_lent={lent} floor_name={floor_name}></Carousel>
         );
       }
       return list;
@@ -196,10 +105,10 @@ export default function Lent(){
                 <nav>
                   <div className="nav nav-tabs border-bottom-0" id="nav-tabs" role="tablist">{navTabs()}</div>
                 </nav>
-            {/* <div className="card"> */}
                 <div className="tab-content" id="nav-tabContent">{navContent()}</div>
                 {/* </div> */}
             </div>
+            <LentModal></LentModal>
         </div>
     );
 }
